Move the ScopeAwareTrait into the framework namespace.
